cleaned up code and fixed memory leak
The memory leak was caused by creating an instance of an object,
var aobj = obj.instance()
Then making a duplicate,
var b = a.duplicate()
add_child(b)
When the func exits the reference to aobj is lost and is orphaned.
To solve this you must  aobj.queue_free() before losing scope (eg. the func exits.